style(lint): zera o flutter analyze para o CI estrito (very_good_analysis)
O CI mobile roda `flutter analyze` puro, que é fatal em qualquer issue (inclusive infos).
Havia 95: 2 warnings reais + 93 infos de estilo.

- Warnings (código morto meu do E2): remove o widget _UfAutocomplete não referenciado em
  customer_register_page (e o parâmetro key órfão).
- Infos: `dart fix --apply` resolveu 70 em 26 arquivos (const, type annotations, ordering,
  sort_constructors, cascades simples, etc.). Os ~8 restantes, manualmente:
  * catch sem `on` → `on Exception` (Dio/repo) e `on Object` nos dois pontos de geocoding
    (o plugin lança tipos fora de Exception em teste — preserva o catch-all original);
  * WillPopScope → PopScope (onPopInvokedWithResult);
  * BitmapDescriptor.fromBytes → .bytes (dpr capturado antes dos awaits p/ não disparar
    use_build_context_synchronously);
  * unawaited() no marker assíncrono do mapa; guarda context.mounted no snackbar de checkout.

Verificação: flutter analyze "No issues found!" + flutter test 151/151.

+++ lib/features/inventory/presentation/pages/inventory_page.dart
@@ -257,7 +257,7 @@ class _InventoryView extends StatelessWidget {
                                         'productName': product.name,
                                         'unit': product.unit,
                                         'currentStock': product.stock
-                                            .toDouble(),
+                                            ,
                                       },
                                     )
                                     .then((result) {
